@@ -13,45 +13,25 @@ import (
 	"github.com/open-policy-agent/regal/internal/util"
 )
 
-// UnquotedPath returns a slice of strings from a path without quotes.
-// e.g. data.foo["bar"] -> ["foo", "bar"], note that the data is not included.
-func UnquotedPath(path ast.Ref) []string {
-	ret := make([]string, 0, len(path)-1)
-	for _, ref := range path[1:] {
-		ret = append(ret, strings.Trim(ref.String(), `"`))
-	}
-
-	return ret
-}
-
 // IsBodyGenerated checks if the body of a rule is generated by the parser.
 func IsBodyGenerated(rule *ast.Rule) bool {
-	if rule.Default {
+	if rule.Default || len(rule.Body) == 0 {
 		return true
 	}
 
-	if len(rule.Body) == 0 {
-		return true
-	}
+	if rule.Head != nil && rule.Body[0] != nil {
+		bodyStart := rule.Body[0].Location
+		if bodyStart == rule.Location || rule.Head.Value != nil && bodyStart == rule.Head.Value.Location {
+			return true
+		}
 
-	if rule.Head == nil {
-		return false
-	}
-
-	if rule.Body[0] != nil && rule.Body[0].Location == rule.Location {
-		return true
-	}
-
-	if rule.Body[0] != nil && rule.Head.Value != nil && rule.Body[0].Location == rule.Head.Value.Location {
-		return true
-	}
-
-	if rule.Head.Key != nil &&
-		rule.Body[0].Location.Row == rule.Head.Key.Location.Row &&
-		rule.Body[0].Location.Col < rule.Head.Key.Location.Col {
-		// This is a quirk in the original AST — the generated body will have a location
-		// set before the key, i.e. "message"
-		return true
+		if rule.Head.Key != nil &&
+			rule.Body[0].Location.Row == rule.Head.Key.Location.Row &&
+			rule.Body[0].Location.Col < rule.Head.Key.Location.Col {
+			// This is a quirk in the original AST — the generated body will have a location
+			// set before the key, i.e. "message"
+			return true
+		}
 	}
 
 	return false
@@ -78,24 +58,12 @@ func RefStringToBody(path string) ast.Body {
 // efficient way possible, using interned terms where possible. See RefStringToBody for
 // more details on the limitations of this function.
 func RefStringToRef(path string) ast.Ref {
-	var i int
-	if i = strings.Index(path, "."); i == -1 {
-		return ast.Ref([]*ast.Term{refHeadTerm(path)})
-	}
+	before, after, found := strings.Cut(path, ".")
+	terms := append(make([]*ast.Term, 0, strings.Count(path, ".")+1), refHeadTerm(before))
 
-	terms := append(make([]*ast.Term, 0, strings.Count(path, ".")+1), refHeadTerm(path[:i]))
-
-	for {
-		path = path[i+1:]
-		if i = strings.Index(path, "."); i == -1 {
-			if len(path) > 0 {
-				terms = append(terms, ast.InternedTerm(path))
-			}
-
-			break
-		}
-
-		terms = append(terms, ast.InternedTerm(path[:i]))
+	for found {
+		before, after, found = strings.Cut(after, ".")
+		terms = append(terms, ast.InternedTerm(before))
 	}
 
 	return ast.Ref(terms)
@@ -149,7 +117,7 @@ func StructToValue(input any) ast.Value {
 			parts := strings.Split(tag, ",")
 			tag = parts[0]
 
-			if slices.Contains(parts[1:], "omitempty") && isZeroValue(value) {
+			if slices.Contains(parts[1:], "omitempty") && value.IsZero() {
 				continue
 			}
 		}
@@ -158,35 +126,6 @@ func StructToValue(input any) ast.Value {
 	}
 
 	return ast.NewObject(kvs...)
-}
-
-func isZeroValue(v reflect.Value) bool {
-	switch v.Kind() {
-	case reflect.Ptr, reflect.Interface:
-		return v.IsNil()
-	case reflect.Slice, reflect.Map, reflect.Array:
-		return v.IsNil() || v.Len() == 0
-	case reflect.String:
-		return v.Len() == 0
-	case reflect.Bool:
-		return !v.Bool()
-	case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-		return v.Int() == 0
-	case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
-		return v.Uint() == 0
-	case reflect.Float32, reflect.Float64:
-		return v.Float() == 0
-	case reflect.Struct:
-		for i := range v.NumField() {
-			if !isZeroValue(v.Field(i)) {
-				return false
-			}
-		}
-
-		return true
-	}
-
-	return false
 }
 
 func toAstValue(v any) ast.Value {
